[DEV-37] rename: 파일명 카멜 케이스에서 케밥 케이스로 수정
